Normalized impulse response
